requirements.txt
numpy==1.26.4
scikit-learn==1.2.2
tensorflow==2.18.0
opencv-python
pillow
joblib
streamlit
xgboost
